+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - `@qiskit/qiskit-sim`: Add print circuit method
 - `@qiskit/qiskit-sim`: Add print method to Circuit API documentation
+- `@qiskit/qiskit-sim`: Add 'add' method to Circuit class
 
 ## [0.8.0] - 2019-04-25